EN-865 Update services and controllers
