+++ engine/src/main/java/org/entando/entando/aps/system/services/entity/IEntityManagerService.java
@@ -14,7 +14,7 @@
 package org.entando.entando.aps.system.services.entity;
 
 import org.entando.entando.aps.system.services.entity.model.EntityManagerDto;
-import org.entando.entando.aps.system.services.entity.model.EntityTypeDto;
+import org.entando.entando.aps.system.services.entity.model.EntityTypeShortDto;
 import org.entando.entando.web.common.model.PagedMetadata;
 import org.entando.entando.web.common.model.RestListRequest;
 
@@ -27,6 +27,6 @@ public interface IEntityManagerService {
 
     public EntityManagerDto getEntityManager(String entityManagerCode);
 
-    public PagedMetadata<EntityTypeDto> getEntityTypes(String entityManagerCode, RestListRequest requestList);
+    public PagedMetadata<EntityTypeShortDto> getShortEntityTypes(String entityManagerCode, RestListRequest requestList);
 
 }
